MiniRogue Game
This game is currently run in linux command line to show the skeleton
and game play as a hole. It will be updated with a GUI on android
platform as a more decent game.
More about this game:
1. I Implemented a game of dungeon adventure in randomly generated 2D
maps.
2. Multiple interactive object classes differed in behaviors or
functionalities were implemented using OOP: 4 categories of monsters, 4
categories of weapons, 5 categories of scrolls, etc.
3. Multiple player operations supported: (1) Make movement; (2) Attack
and choose attack subject; (3) Pick up items; (4) See inventory; (5) Use
items in inventory; (6) Cheat to have overwhelming power.
4. Smart monster behavior realized by using inheritance and BFS
algorithm.